one hot encode/decode~

+++ supervised_learning/0x01-classification/24-one_hot_encode.py
@@ -14,4 +14,4 @@ def one_hot_encode(Y, classes):
        or len(Y.shape) != 1\
        or classes != Y.max()+1:
         return None
-    return np.eye(classes)[Y].T.astype(int)
+    return np.eye(classes)[Y].T
